Back-end/requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
django-cors-headers==4.3.1
django-optimized-image==0.3.0
djangorestframework==3.15.1
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
packaging==23.2
pillow==10.3.0
pluggy==1.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
Pygments==2.17.2
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.0
SQLAlchemy==2.0.27
sqlparse==0.4.4
stripe==9.1.0
tinify==1.6.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
